Rename namespaces to SprykerSmartInboxConnector
