+++ .gitignore
@@ -2,4 +2,4 @@ composer.phar
 /vendor/
 .gitignore
 composer.lock
-tests/OneAndOneTest/Zed/OneAndOneMailConnector/_support/_generated
+tests/FlagbitTest/Zed/SprykerSmartInboxConnector/_support/_generated
