+++ src/Flagbit/Shared/SprykerSmartInboxConnector/OneAndOneMailConnectorConstants.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Flagbit\Shared\SprykerSmartInboxConnector;
+
+interface OneAndOneMailConnectorConstants
+{
+    public const MATRIX_KEY = 'ONE_AND_ONE_MAIL_CONNECTOR_MATRIX_KEY';
+    public const SHOP_NAME = 'ONE_AND_ONE_MAIL_CONNECTOR_SHOP_NAME';
+}
