+++ src/Flagbit/Shared/SprykerSmartInboxConnector/Transfer/mail.transfer.xml
