+++ src/Flagbit/Zed/SprykerSmartInboxConnector/Communication/navigation.xml
