+++ src/Flagbit/Zed/SprykerSmartInboxConnector/Persistence/OneAndOneMailConnectorRepository.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Flagbit\Zed\SprykerSmartInboxConnector\Persistence;
+
+use Propel\Runtime\ActiveQuery\Criteria;
+use Spryker\Zed\Kernel\Persistence\AbstractRepository;
+
+/**
+ * @method \Flagbit\Zed\SprykerSmartInboxConnector\Persistence\OneAndOneMailConnectorPersistenceFactory getFactory()
+ */
+class OneAndOneMailConnectorRepository extends AbstractRepository implements OneAndOneMailConnectorRepositoryInterface
+{
+    /**
+     * @param array $idOrderItems
+     *
+     * @return mixed
+     */
+    public function findSpySalesOrderItemByIdWithLastStateChange(array $idOrderItems)
+    {
+        return $this
+            ->getFactory()
+            ->createSpySalesOrderItemQuery()
+            ->orderByLastStateChange(Criteria::DESC)
+            ->filterByIdSalesOrderItem($idOrderItems, Criteria::IN)
+            ->findOne();
+    }
+}
