+++ src/Flagbit/Zed/SprykerSmartInboxConnector/Persistence/OneAndOneMailConnectorRepositoryInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Flagbit\Zed\SprykerSmartInboxConnector\Persistence;
+
+interface OneAndOneMailConnectorRepositoryInterface
+{
+    /**
+     * @param array $idOrderItems
+     *
+     * @return mixed
+     */
+    public function findSpySalesOrderItemByIdWithLastStateChange(array $idOrderItems);
+}
