+++ src/Flagbit/Zed/SprykerSmartInboxConnector/Presentation/Index/index.twig
