--- src/OneAndOne/Shared/OneAndOneMailConnector/OneAndOneMailConnectorConstants.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace OneAndOne\Shared\OneAndOneMailConnector;
-
-interface OneAndOneMailConnectorConstants
-{
-    public const MATRIX_KEY = 'ONE_AND_ONE_MAIL_CONNECTOR_MATRIX_KEY';
-    public const SHOP_NAME = 'ONE_AND_ONE_MAIL_CONNECTOR_SHOP_NAME';
-}
