--- src/OneAndOne/Zed/OneAndOneMailConnector/Business/OneAndOneMailConnectorFacadeInterface.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace OneAndOne\Zed\OneAndOneMailConnector\Business;
-
-use Generated\Shared\Transfer\MailTransfer;
-use Generated\Shared\Transfer\OrderTransfer;
-
-interface OneAndOneMailConnectorFacadeInterface
-{
-    /**
-     * @param \Generated\Shared\Transfer\MailTransfer $mailTransfer
-     * @param \Generated\Shared\Transfer\OrderTransfer $orderTransfer
-     *
-     * @return \Generated\Shared\Transfer\MailTransfer
-     */
-    public function expandOrderMailTransfer(MailTransfer $mailTransfer, OrderTransfer $orderTransfer): MailTransfer;
-}
