--- src/OneAndOne/Zed/OneAndOneMailConnector/Business/ParcelDelivery/ParcelDeliveryFactory.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace OneAndOne\Zed\OneAndOneMailConnector\Business\ParcelDelivery;
-
-use Generated\Shared\Transfer\ParcelDeliveryTransfer;
-
-class ParcelDeliveryFactory
-{
-    /**
-     * @return \Generated\Shared\Transfer\ParcelDeliveryTransfer
-     */
-    public function create(): ParcelDeliveryTransfer
-    {
-        return new ParcelDeliveryTransfer();
-    }
-}
