--- src/OneAndOne/Zed/OneAndOneMailConnector/Business/SchemaOrgOrderMailExpander/SchemaOrgOrderMailExpanderInterface.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace OneAndOne\Zed\OneAndOneMailConnector\Business\SchemaOrgOrderMailExpander;
-
-use Generated\Shared\Transfer\MailTemplateTransfer;
-use Generated\Shared\Transfer\MailTransfer;
-use Generated\Shared\Transfer\OrderTransfer;
-use Generated\Shared\Transfer\SchemaOrgTransfer;
-
-interface SchemaOrgOrderMailExpanderInterface
-{
-    /**
-     * Expand the MailTransfer with an MailTemplateTransfer which adds additional schema.org information.
-     *
-     * @param \Generated\Shared\Transfer\MailTransfer $mailTransfer
-     * @param \Generated\Shared\Transfer\OrderTransfer $orderTransfer
-     * @param \Generated\Shared\Transfer\MailTemplateTransfer $mailTemplateTransfer
-     * @param \Generated\Shared\Transfer\SchemaOrgTransfer $schemaOrgTransfer
-     *
-     * @return \Generated\Shared\Transfer\MailTransfer
-     */
-    public function expandOrderMailTransfer(
-        MailTransfer $mailTransfer,
-        OrderTransfer $orderTransfer,
-        MailTemplateTransfer $mailTemplateTransfer,
-        SchemaOrgTransfer $schemaOrgTransfer
-    ): MailTransfer;
-}
