--- src/OneAndOne/Zed/OneAndOneMailConnector/Persistence/OneAndOneMailConnectorPersistenceFactory.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace OneAndOne\Zed\OneAndOneMailConnector\Persistence;
-
-use Orm\Zed\Sales\Persistence\SpySalesOrderItemQuery;
-use Spryker\Zed\Kernel\Persistence\AbstractPersistenceFactory;
-
-/**
- * @method \OneAndOne\Zed\OneAndOneMailConnector\OneAndOneMailConnectorConfig getConfig()
- * @method \OneAndOne\Zed\OneAndOneMailConnector\Persistence\OneAndOneMailConnectorRepositoryInterface getRepository()
- */
-class OneAndOneMailConnectorPersistenceFactory extends AbstractPersistenceFactory
-{
-    /**
-     * @return \Orm\Zed\Sales\Persistence\SpySalesOrderItemQuery
-     */
-    public function createSpySalesOrderItemQuery(): SpySalesOrderItemQuery
-    {
-        return SpySalesOrderItemQuery::create();
-    }
-}
