--- src/OneAndOne/Zed/OneAndOneMailConnector/Persistence/OneAndOneMailConnectorRepositoryInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace OneAndOne\Zed\OneAndOneMailConnector\Persistence;
-
-interface OneAndOneMailConnectorRepositoryInterface
-{
-    /**
-     * @param array $idOrderItems
-     *
-     * @return mixed
-     */
-    public function findSpySalesOrderItemByIdWithLastStateChange(array $idOrderItems);
-}
