--- tests/OneAndOneTest/Zed/OneAndOneMailConnector/codeception.yml
@@ -1,24 +1,0 @@
-namespace: OneAndOneTest\Zed\OneAndOneMailConnector
-
-paths:
-  tests: .
-  data: ../../../_data
-  support: _support
-  log: ../../../_output
-  output: ../../../_output
-
-coverage:
-  enabled: true
-  remote: false
-  whitelist: { include: ['../../../../src/*'] }
-
-suites:
-  Business:
-    path: Business
-    class_name: OneAndOneMailConnectorBusinessTester
-    modules:
-      enabled:
-        - Asserts
-        - \SprykerTest\Shared\Testify\Helper\Environment
-        - \SprykerTest\Shared\Testify\Helper\LocatorHelper:
-            projectNamespaces: ['OneAndOne']
